# OTT-Platform-Web-Application/Backend/docs/api_docs.yaml
...
openapi: 3.0.3
info:
  title: Dhaakad API
  description: Dhaakad API documentation (Might be buggy)
  version: 0.0.5
  contact:
    email: vishal.rustagi@metaorangedigital.com
servers:
  - url: 'https://api.dhaakadcinema.com/api/v1'
...
paths:
...
  /reviews:
    get:
      tags:
        - reviews
      summary: get all reviews
      operationId: getAllReviews
      responses:
        200:
          description: fetch successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "fetch successfully"
                  results:
                    type: integer
                    example: 10
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - reviews
      summary: add review
      operationId: addReview
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                review:
                  type: string
                  example: 'Good Movie'
                description:
                  type: string
                  example: 'Good Movie'
                rating:
                  type: integer
                  example: 1||2||3||4||5
                movieId:
                  type: string
                  example: 'asdasd'
                user:
                  type: string
                  example: 'abc123'
      responses:
        201:
          description: created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "created successfully"
                  data:
                    type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
      security:
        - bearerAuth: [ ]
...
  /reviews/updatestatus:
    post:
      tags:
        - reviews
      summary: Update Status
      operationId: updateStatus
      parameters:
        - name: reviewId
          in: query
          required: true
          schema:
            type: string
            example: xyz123
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  example: 'abc123'
                review:
                  type: string
                  example: 'Good Movie'
                description:
                  type: string
                  example: 'Good Movie'
                rating:
                  type: integer
                  example: 1||2||3||4||5
                movieId:
                  type: string
                  example: 'asdasd'
                isApproved:
                  type: boolean
                  example: true
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "updated successfully"
                  data:
                    type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: [ ]
...
  /rewards:
    get:
      tags:
        - rewards
      summary: Get all rewards
      operationId: getAllRewards
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Rewards fetched successfully"
                  results:
                    type: integer
                    example: 10
                  data:
                    type: object
                    properties:
                      reward:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reward'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - rewards
      summary: Add reward
      operationId: addReward
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 'New Subscription'
                subtitle:
                  type: string
                  example: 'English'
                couponCode:
                  type: string
                  example: 'asd'
                amount:
                  type: integer
                  example: 100
                description:
                  type: string
                  example: 'asdasd'
                expireDate:
                  type: string
                  format: date
                  example: '2024-08-12'
                userLimit:
                  type: string
                  example: 4
                status:
                  type: string
                  example: 'Active||Inactive'
                movieId:
                  type: string
                  example: 'asdasd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Rewards created successfully"
                  data:
                    type: object
                    properties:
                      reward:
                        $ref: '#/components/schemas/Reward'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /notifications/:
    get:
      tags:
        - notifications
      operationId: getAllNotifications
      summary: Get all notifications
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  message:
                    type: string
                    example: 'Notification fetched successfully'
                  results:
                    type: number
                    example: 10
                  data:
                    type: object
                    properties:
                      notification:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - notifications
      operationId: createNotification
      summary: To create a notification
      requestBody:
        description: notification body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  message:
                    type: string
                    example: 'created successfully!!'
                  data:
                    type: object
                    properties:
                      notification:
                        $ref: '#/components/schemas/Notification'
      security:
        - bearerAuth: [ ]
  /trasncations/createOrder:
    post:
      tags:
        - transactions
      summary: To create movie purchase order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - movieId
                - amount
                - currency
                - receipt
              properties:
                userId:
                  type: string
                  example: 'askjdn123'
                movieId:
                  type: string
                  example: 'askjdn1233'
                amount:
                  type: number
                  example: 3654
                currency:
                  type: string
                  example: 'INR'
                receipt:
                  type: string
                  example: '123123asd'
                notes:
                  type: object
                streamed:
                  type: boolean
                  default: false
                  example: false
      responses:
        200:
          description: order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Order Created Succcessfully!"
                  results:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/RazorpayOrder'
...
  /purchase/:
    get:
      tags:
        - purchase
      operationId: getAllPurchase
      summary: Get all purchase
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  message:
                    type: string
                    example: 'PurchaseMovie fetched successfully'
                  results:
                    type: number
                    example: 10
                  data:
                    type: object
                    properties:
                      purchase:
                        type: array
                        items:
                          $ref: '#/components/schemas/Purchase'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - purchase
      operationId: createPurchase
      summary: To create a purchase
      requestBody:
        description: purchase body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  message:
                    type: string
                    example: 'created successfully!!'
                  data:
                    type: object
                    properties:
                      purchase:
                        $ref: '#/components/schemas/Purchase'
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    Review:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'xyz123'
        review:
          type: string
          example: 'Good Movie'
        description:
          type: string
          example: 'Good Movie'
        rating:
          type: integer
          example: 1||2||3||4||5
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        movieId:
          type: string
          example: 'asdasd'
        user:
          type: string
          example: 'abc123'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        isApproved:
          type: boolean
          example: false
...
    Reward:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'pqr123'
        title:
          type: string
          example: 'New Subscription'
        subtitle:
          type: string
          example: 'English'
        couponCode:
          type: string
          example: 'asd'
        amount:
          type: integer
          example: 100
        description:
          type: string
          example: 'asdasd'
        expireDate:
          type: string
          format: date
          example: '2024-08-12'
        userLimit:
          type: string
          example: 4
        status:
          type: string
          example: 'Active||Inactive'
        movieId:
          type: string
          example: 'asdasd'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'sdfsf'
        title:
          type: string
          example: 'Test notification'
        description:
          type: string
          example: 'Test notification description'
        isViewed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    Purchase:
      type: object
      properties:
        userId:
          type: string
          example: 'kjandjk123123'
        movieId:
          type: string
          example: 'askdq213123'
        transitionId:
          type: string
          example: 'alskdmalskdm'
        stream:
          type: boolean
          default: false
          example: false
        expiredAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    RazorpayOrder:
      type: object
      properties:
        id:
          type: string
          example: 'ksjdn234'
        entity:
          type: string
          example: 'Sample'
        amount_paid:
          type: number
          example: 235
        amount_due:
          type: number
          example: 256
        status:
          type: string
          enum:
            - created
            - attempted
            - paid
        attempts:
          type: number
          example: 2
        created_at:
          type: number
          example: 12783123
        description:
          type: string
          example: 'Test description'
        payments:
          type: object
        offers:
          type: object
        transfers:
          type: object
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: 'Something went wrong'
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes